fifth commit
Typography basically deals with how text is formatted on a webpage. Key concepts include importing custom fonts and font stacking. Font stacking involves use of fall back fonts that are alternately loaded (temporarily downloaded) by a user's browser when they load the page until one that works is found. the final fall back font is the sans-serif font family which is the default browser font. another is the inline span selector which allows application of a specific style to parts of inline text e.g. bold a certain letter in a sentence.